Basic audio cropping and playback functionality
Still a work in progress, interface is excessively basic using getchar() and
commandline arguments need to be changed to handle just being given an input
filename, either through output to stdout or asking for a filename. If given
<input filename> <output filename> <start time> <stop time> it should handle
everything without the playback and interface.

crop currently allows for adjusting the crop window
on a second by second basis using a/f and h/l to adjust the start and end.
p will start/stop playback and q quits.